// src/modules/@library/constant/enums.ts
export enum BookTypeEnum {
  Book = "Book",
  Ebook = "Ebook",
  Book_Ebook = "Book_Ebook",
}

export enum BookFileTypeEnum {
  Audio = "Audio",
  Video = "Video",
  PDF = "PDF",
  PPT = "PPT",
}

export enum FineStatusEnum {
  Pending = "Pending",
  Paid = "Paid",
  Discard = "Discard",
}

export enum PatronTypeEnum {
  Student = "Student",
  Staff = "Staff",
  Other = "Other",
}

export enum BookStatusEnum {
  Checkout = "Checkout",
  Checkin = "Checkin",
  Missing = "Missing",
}

export enum LibraryLanguageEnum {
  English = "English",
  Nepali = "Nepali",
  Hindi = "Hindi",
  Sanskrit = "Sanskrit",
  Other = "Other",
}

export enum LibraryBookCategoryEnum {
  Arts_Music = "Arts & Music",
  Business = "Business",
  Cooking = "Cooking",
  History = "History",
  Other = "Other",
}

export enum BookCopyStatusEnum {
  Checkout = "Checkout",
  Available = "Available",
  Missing = "Missing",
  Lost = "Lost",
  Discarded = "Discarded",
}

export enum LanguageEnum {
  English = "English",
  Nepali = "Nepali",
  Hindi = "Hindi",
  Sanskrit = "Sanskrit",
  Other = "Other",
}

export enum FineStatusEnum {
  PENDING = "Pending",
  PAID = "Paid",
  DISCARD = "Discard",
}